Move product-detail-card to product-detail-dir
