+++ projects/packages/my-jetpack/_inc/components/product-detail/stories/mock-data.js
