+++ projects/packages/my-jetpack/_inc/components/product-detail/style.module.scss
@@ -1,23 +1,9 @@
-.card {
-	// Card main wrapper
-	--product-card-bg-color: white;
-	--product-card-border-color: #d5d5d5;
-	--product-card-shadow: rgb(0 0 0 / 3%);
-	--product-card-border-radius: 4px;
-	
-	background-color: var(--product-card-bg-color);
-	border: 1px solid var(--product-card-border-color);
-	border-radius: var(--product-card-border-radius);
-	box-shadow: 0 2px 6px var(--product-card-shadow), 0 1px 2px var(--product-card-shadow);
-	max-width: 600px;
-}
-
 .container {
 	// @todo find a better way to change link color
 	// since this could affect more places
 	--wp-admin-theme-color: var(--jp-black);
 
-	// all css variables used down in product-detail-card
+	// all css variables used down in product-detail
 
 	// Colors
 	--jp-green-primary: #069e08;
